refactor: move load_scores_loss_signed into data
It is a thin sign-normalising wrapper over load_scores, which lives in
data alongside the store format it reads — Scores, offsets, to_grid — so
the two belong together. Nothing about it is validation-specific; magic
imported it from validate only because that is where it happened to sit.

No new dependency edge: data already imports from config.config, and the
config layer does not import data, so ScoreConfig and load_subconfig come
along without a cycle.
