+++ docs/magic.rst
@@ -35,7 +35,7 @@ After a run completes, ``run_cfg.run_path`` contains:
 * ``scores/`` — a score directory, the same self-describing format the
   scoring pipeline writes. ``info.json`` records ``attribute_tokens`` and
   ``num_scores``, so consumers never infer the layout from the shape.
-  Read it with :func:`bergson.validate.load_scores_loss_signed`, which
+  Read it with :func:`bergson.data.load_scores_loss_signed`, which
   returns ``(scores, multi_query)``:
 
   * Per-example: ``(num_train_docs, 1)``, indexed directly by ``doc_id``.
@@ -57,7 +57,7 @@ After a run completes, ``run_cfg.run_path`` contains:
 
   .. code-block:: python
 
-     from bergson.validate import load_scores_loss_signed
+     from bergson.data import load_scores_loss_signed
 
      scores, _ = load_scores_loss_signed("runs/magic/scores")
      doc_ids = torch.from_numpy(np.load("runs/magic/scores/doc_ids.npy"))
